wip: rename and update `getClangFormatPath.js`
